#60 Cleanup code and refactor to support Box in Job class
