+++ src/main/java/com/clayfactoria/codecs/task/TaskExecutor.java
@@ -1,0 +1,16 @@
+package com.clayfactoria.codecs.task;
+
+import com.hypixel.hytale.component.Ref;
+import com.hypixel.hytale.math.shape.Box;
+import com.hypixel.hytale.math.vector.Vector3d;
+import com.hypixel.hytale.server.core.universe.world.World;
+import com.hypixel.hytale.server.core.universe.world.storage.EntityStore;
+
+public interface TaskExecutor {
+  boolean canPerformTask(Ref<EntityStore> entityRef);
+  boolean execute(Ref<EntityStore> entityRef);
+
+  boolean usesBounds();
+  void checkBounds(Box bounds) throws IllegalArgumentException;
+  Vector3d findNextWalkLocationInBounds(Box bounds, World world, Vector3d from);
+}
